[UPDATE] 페이지네이션 방식 수정
기존 offset, size 방식으로 페이징 처리했으나 성능상 좋지 않았기 때문에 커서 방식으로 변경 (커서 기준은 pk 값으로 처리)

+++ src/test/java/com/ssg/usms/business/accident/repository/AccidentRepositoryTest.java
@@ -104,7 +104,7 @@ public class AccidentRepositoryTest {
         Long storeId = 1L;
         long startTimestamp = 1000000000L;
         long endTimestamp = System.currentTimeMillis();
-        int offset = 0;
+        long offset = 0L;
         int size = 20;
         List<AccidentBehavior> behaviors = new ArrayList<>();
         behaviors.add(AccidentBehavior.FIGHT);
@@ -127,7 +127,7 @@ public class AccidentRepositoryTest {
         Long storeId = 1L;
         long startTimestamp = 1000000000L;
         long endTimestamp = System.currentTimeMillis();
-        int offset = 0;
+        long offset = 0L;
         int size = 20;
 
         //when
@@ -141,7 +141,7 @@ public class AccidentRepositoryTest {
     public void testFindAccidentRegion() {
 
         //given
-        int offset = 0;
+        long offset = 0L;
         int size = 20;
         long startTimestamp = 0;
         long endTimestamp = System.currentTimeMillis();
